+++ backend/controllers/SiteController.php
@@ -1,6 +1,7 @@
 <?php
 namespace backend\controllers;
 
+use common\components\keyStorage\FormModel;
 use Yii;
 
 /**
@@ -24,5 +25,55 @@ class SiteController extends \yii\web\Controller
     {
         $this->layout = Yii::$app->user->isGuest ? 'base' : 'common';
         return parent::beforeAction($action);
+    }
+
+    public function actionSettings()
+    {
+        $model = new FormModel([
+            'keys' => [
+                'frontend.maintenance' => [
+                    'label' => Yii::t('backend', 'Frontend maintenance mode'),
+                    'type' => FormModel::TYPE_DROPDOWN,
+                    'items' => [
+                        0 => Yii::t('backend', 'Disabled'),
+                        1 => Yii::t('backend', 'Enabled')
+                    ]
+                ],
+                'backend.theme-skin' => [
+                    'label' => Yii::t('backend', 'Backend theme'),
+                    'type' => FormModel::TYPE_DROPDOWN,
+                    'items' => [
+                        'skin-black' => 'skin-black',
+                        'skin-blue' => 'skin-blue',
+                        'skin-green' => 'skin-green',
+                        'skin-purple' => 'skin-purple',
+                        'skin-red' => 'skin-red',
+                        'skin-yellow' => 'skin-yellow'
+                    ]
+                ],
+                'backend.layout-fixed' => [
+                    'label' => Yii::t('backend', 'Fixed backend layout'),
+                    'type' => FormModel::TYPE_CHECKBOX
+                ],
+                'backend.layout-boxed' => [
+                    'label' => Yii::t('backend', 'Boxed backend layout'),
+                    'type' => FormModel::TYPE_CHECKBOX
+                ],
+                'backend.layout-collapsed-sidebar' => [
+                    'label' => Yii::t('backend', 'Backend sidebar collapsed'),
+                    'type' => FormModel::TYPE_CHECKBOX
+                ]
+            ]
+        ]);
+
+        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            Yii::$app->session->setFlash('alert', [
+                'body' => Yii::t('frontend', 'Settings was successfully saved'),
+                'options' => ['class' => 'alert alert-success']
+            ]);
+            return $this->refresh();
+        }
+
+        return $this->render('settings', ['model' => $model]);
     }
 }